% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEllipses.R
\name{calculateEllipses}
\alias{calculateEllipses}
\title{calculateEllipses
calculates the coordinates of confidence ellipses}
\usage{
calculateEllipses(
  suppIndividualTable,
  vep,
  axes = c(1, 2),
  confInt = 0.9,
  ellipseType = "barycentric",
  productName = "product",
  subjectName = "subject",
  ellipseCalculation = "Chi",
  bootstrap = FALSE,
  nSamples = 100
)
}
\arguments{
\item{suppIndividualTable}{an extended data frame with "product","subject" and as many further column as variables (descriptors) in the dataframe}

\item{vep}{matrix of eigenvectors obtained by PCA or CVA}

\item{axes}{the two axes to be chosen for the plot}

\item{confInt}{threshold for confidence intervall. A value of 0.9 (the default) for barycentric ellipses indicates that the mean is inside the ellipse at 90\%. For individual ellipses, one individual has 90\% of chances to be in the ellipse}

\item{ellipseType}{"barycentric" or "individual". Barycentric is the confidence ellipse for the mean while individual is the confidence ellipse for the individuals.}

\item{productName}{name of the column containing the main effect. "product" by default.}

\item{subjectName}{name of the column containing the variability around the main effect. "subject" by default.}

\item{ellipseCalculation}{"Chi","F" or "SAS". Method to estimate the ellipse calculation.}

\item{bootstrap}{=FALSE If TRUE the ellipses represents the variability}

\item{nSamples}{=100 Numer of samples used with bootstrap option}
}
\description{
calculateEllipses
calculates the coordinates of confidence ellipses
}
\details{
If missing data
}
\examples{
{
# df=data.frame(product=rep(c("a","b","c"),each=10),
# subject=rep(paste0("S",1:10),3),acide=rnorm(30),amer=rnorm(30))
# respca=PCA(df)
# calculateEllipses(df,respca$EigenVectors)
}
}
