% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanString.R
\name{cleanString}
\alias{cleanString}
\title{Clean a string.}
\usage{
cleanString(x, toLower = FALSE)
}
\arguments{
\item{x}{A string.}

\item{toLower}{If TRUE, the string is transformed to lower case.}
}
\value{
The cleaned string.
}
\description{
Clean a string (removal of special characters and accented characters, removal of multiple spaces).
}
\keyword{internal}
