% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.R
\name{CVA}
\alias{CVA}
\title{CVA}
\usage{
CVA(
  extendedData,
  test = "Hotelling-Lawley",
  nbDimHotelling = NULL,
  option = "TwoWayANOVA",
  representation = "Biplot"
)
}
\arguments{
\item{extendedData}{an extended dataframe whose first three columns are "subject", "product" and "rep". All the other columns are variables (as many columns as descriptors for sensory analysis)}

\item{test}{"Pillai", "Wilks", "Hotelling-Lawley" or "Roy". Default to "Hotelling-Lawley". The name of the test statistic to be used. Partial matching is used so the name can be abbreviated.}

\item{nbDimHotelling}{If NULL, nbDimSig,}

\item{option}{"OneWayANOVA","TwoWayANOVA","MAM","MultiMAM" or "TwoWayWithoutTakingSubjectEffectIntoEllipses". The type of model chosen for the CVA}

\item{representation}{"Biplot" or "TwoMaps"}
}
\value{
a list containing IndivCoord,VarCoord,IndSup,EigenVectors, EigenValues
}
\description{
Canonical Variate Analysis for sensory data
}
