% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVAgg.r
\name{CVAgg}
\alias{CVAgg}
\title{CVAgg
Calculates the CVAs (potentially by period of time) and returns coordinates as long data.frames to make easier ggplot}
\usage{
CVAgg(
  df,
  confInt = 0.9,
  axes = list(c(1, 2)),
  option = "TwoWayANOVA",
  representation = "DistanceBiplot",
  ellipseType = "barycentric",
  productName = "product",
  subjectName = "subject",
  ellipseCalculation = "Chi",
  bootstrap = FALSE,
  nSamples = 100,
  nbDimHotelling = NULL,
  ...
)
}
\arguments{
\item{df}{a long dataframe with subject, product, descriptor, score, rep and period}

\item{confInt}{threshold for confidence intervall. A value of 0.9 (the default) for barycentric ellipses indicates that the mean is inside the ellipse at 90\%. For individual ellipses, one individual has 90\% of chances to be in the ellipse}

\item{axes}{the two axes to be chosen for the plot}

\item{option}{"OneWayANOVA","TwoWayANOVA","MAM","MultiMAM" or "TwoWayWithoutTakingSubjectEffectIntoEllipses". The type of model chosen for the CVA}

\item{representation}{"Biplot" or "TwoMaps"}

\item{ellipseType}{"barycentric" or "individual". Barycentric is the confidence ellipse for the mean while individual is the confidence ellipse for the individuals.}

\item{productName}{name of the column containing the main effect. "product" by default.}

\item{subjectName}{name of the column containing the variability around the main effect. "subject" by default.}

\item{ellipseCalculation}{"Chi","F" or "SAS". Method to estimate the ellipse calculation.}

\item{bootstrap}{=FALSE If TRUE the ellipses represents the variability}

\item{nSamples}{=100 Numer of samples used with bootstrap option}

\item{nbDimHotelling}{If NULL, nbDimSig,}

\item{...}{Further parameters in PCA and in calculateEllipses}
}
\value{
listCoord
}
\description{
CVAgg
Calculates the CVAs (potentially by period of time) and returns coordinates as long data.frames to make easier ggplot
}
