% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBarPlot.R
\name{gBarPlot}
\alias{gBarPlot}
\title{Plot a barplot.}
\usage{
gBarPlot(
  df,
  y = "score",
  x = "descriptor",
  group = "product",
  fill = "product",
  groupvars = c("product", "rep", "descriptor", "period"),
  confInt = 0.95,
  errorBars = "CI",
  ...
)
}
\arguments{
\item{df}{A dataframe}

\item{y}{Name of numeric variable of df corresponding to the Y axis of the barplot.}

\item{x}{Name of character variable of df corresponding to the X axis of the curve.}

\item{group}{The group variable}

\item{fill}{TODO}

\item{groupvars}{A vector of coulmns to be aggregated as the experimental unit.}

\item{confInt}{The confidence interval for the bars.}

\item{errorBars}{The type of error bar: "", SD, SEM, CI}

\item{...}{List of additional parameters:\cr
- title Character. Main title of the plot.\cr
- facet Character: "" (standard ggplot), "grid" (matrix of panels) or "wrap" (wrap panel)\cr
- rows Character: if facet = "grid", rows of the matrix of panels\cr
- cols Character: if facet = "grid", columns of the matrix of panels\cr
- wrap Character: if facet = "wrap", name of the wraping variable\cr
- labX Character: X lab title\cr
- labY Character: Y lab title\cr
- grid Logical: if TRUE, display grid\cr
- labels Character\cr
- colors Character\cr
- scales Character\cr
- vnames Friendly names of the variable.}
}
\value{
A plot of class ggplot.
}
\description{
Extends gBasePlot.
}
