% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBasePlot.R
\name{gBasePlot}
\alias{gBasePlot}
\title{Base plot.}
\usage{
gBasePlot(...)
}
\arguments{
\item{...}{List of additional parameters:\cr
- title Character. Main title of the plot.\cr
- facet Character: "" (standard ggplot), "grid" (matrix of panels) or "wrap" (wrap panel)\cr
- rows Character: if facet = "grid", rows of the matrix of panels\cr
- cols Character: if facet = "grid", columns of the matrix of panels\cr
- wrap Character: if facet = "wrap", name of the wraping variable\cr
- labX Character: X lab title\cr
- labY Character: Y lab title\cr
- grid Logical: if TRUE, display grid\cr
- labels Character\cr
- colors Character\cr
- scales Character\cr
- vnames Friendly names of the variable.}
}
\value{
An object of class ggplot
}
\description{
Base function for ggplot.
}
\keyword{internal}
