% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDistributionPlot.R
\name{gDistributionPlot}
\alias{gDistributionPlot}
\title{Plot distributions (boxplot, violin).}
\usage{
gDistributionPlot(
  df,
  plot = c("violin", "boxplot"),
  y = "score",
  x = "product",
  fill = "product",
  ...
)
}
\arguments{
\item{df}{Dataframe}

\item{plot}{Character: "violin" and/or "boxplot" and/or "jitter"}

\item{y}{Character. Name of numeric variable of df corresponding to the Y axis of the plot.}

\item{x}{Character. Name of character variable of df corresponding to the Y axis of the curve.}

\item{fill}{Character. Name of character variable of df corresponding to the label of the boxplots/violins/jitters.}

\item{...}{List of additional parameters:\cr
- title Character. Main title of the plot.\cr
- facet Character: "" (standard ggplot), "grid" (matrix of panels) or "wrap" (wrap panel)\cr
- rows Character: if facet = "grid", rows of the matrix of panels\cr
- cols Character: if facet = "grid", columns of the matrix of panels\cr
- wrap Character: if facet = "wrap", name of the wraping variable\cr
- labX Character: X lab title\cr
- labY Character: Y lab title\cr
- grid Logical: if TRUE, display grid\cr
- labels Character\cr
- colors Character\cr
- scales Character\cr
- vnames Friendly names of the variable.}
}
\value{
A plot of class ggplot.
}
\description{
Extends gBasePlot.
}
