% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMapPlot.R
\name{gMapPlot}
\alias{gMapPlot}
\title{Plot a map.}
\usage{
gMapPlot(
  listCoord,
  type = "points",
  output = c("ind", "var"),
  indsup = FALSE,
  ellipses = TRUE,
  ...
)
}
\arguments{
\item{listCoord}{List of dataframes \cr
$indivCoord : dataframe of individual coordinates with columns name/x/y/axes
$varCoord : dataframe of variable coordinates with columns name/x/y/axes
$indivEllipsesCoord : dataframe of ellipse coordinates with columns name/x/y/axes
$inertia : dataframe of explained variance x/y/axes}

\item{type}{Character: "points" or "arrows"}

\item{output}{=c("ind","var")}

\item{indsup}{Boolean indicating whether individuals should be projected on the map or not (FALSE by default)}

\item{ellipses}{Boolean indicating whether confidence ellipses should be on the map or not (TRUE by default)}

\item{...}{List of additional parameters:\cr
- title Character. Main title of the plot.\cr
- facet Character: "" (standard ggplot), "grid" (matrix of panels) or "wrap" (wrap panel)\cr
- rows Character: if facet = "grid", rows of the matrix of panels\cr
- cols Character: if facet = "grid", columns of the matrix of panels\cr
- wrap Character: if facet = "wrap", name of the wraping variable\cr
- labX Character: X lab title\cr
- labY Character: Y lab title\cr
- grid Logical: if TRUE, display grid\cr
- labels Character\cr
- colors Character\cr
- scales Character\cr
- vnames Friendly names of the variable.}
}
\value{
A plot of class ggplot.
}
\description{
Extends gBasePlot.
}
