% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSavePlot.R
\name{gSavePlot}
\alias{gSavePlot}
\title{Save a ggplot.}
\usage{
gSavePlot(gg, format, fileName)
}
\arguments{
\item{gg}{Object of class ggplot}

\item{format}{Format of file: "", "html", eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only)}

\item{fileName}{File name.}
}
\value{
Object of class ggplot or kable
}
\description{
Save a ggplot to an external file.
}
