% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelPerformances.R
\name{PanelPerformances}
\alias{PanelPerformances}
\title{PanelPerformances}
\usage{
PanelPerformances(
  frame,
  modelType = "overall",
  negativeCorrection = TRUE,
  correctOnlyIfSignificant = FALSE,
  limitOfSignificance = 0.05,
  onlySignificantDim = FALSE,
  manovaTest = "Hotelling",
  panelistPerf = FALSE,
  correlationTest = "none",
  multidimLine = FALSE,
  fisherRatio = FALSE,
  levelOption = FALSE,
  whenOverallUseMAMForTest = FALSE
)
}
\arguments{
\item{frame}{a data.frame whose first colnames are subject, product, replicate, att 1, ..., attribute p}

\item{modelType}{"overall" "mam" or "classic"}

\item{negativeCorrection}{boolean. If TRUE, a correction is made for removing negative issue}

\item{correctOnlyIfSignificant}{boolean. If TRUE, the scaling effect is corrected only when it is significant}

\item{limitOfSignificance}{limit ofSignificance. Default to 0.05.}

\item{onlySignificantDim}{boolean}

\item{manovaTest}{"Hotelling"}

\item{panelistPerf}{FALSE}

\item{correlationTest}{"none"}

\item{multidimLine}{FALSE}

\item{fisherRatio}{FALSE}

\item{levelOption}{=FALS}

\item{whenOverallUseMAMForTest}{=FALSE}
}
\description{
PanelPerformances
}
