% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA
Performs a Principal Component Analysis}
\usage{
PCA(
  extendedData,
  dataType = "productMeans",
  option = "covariance",
  representation = "DistanceBiplot"
)
}
\arguments{
\item{extendedData}{an extended dataframe with "subject", "product" as two first columns then a column by variable (descriptor)}

\item{dataType}{"productMeans" or "raw"}

\item{option}{"Covariance" or "Correlations".}

\item{representation}{"DistanceBiplot","CorrelationBiplot" or "TwoMaps".}
}
\value{
A list contaning the PCA results: IndivCoord, VarCoord,IndSup, EigenVectors,EigenValues
}
\description{
PCA
Performs a Principal Component Analysis
}
\examples{
{
df=data.frame(product=rep(c("a","b","c"),each=10),
subject=rep(paste0("S",1:10),3),acide=rnorm(30),amer=rnorm(30))
 respca=PCA(df)
}
}
