% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resPCA.r
\name{plot.resPCA}
\alias{plot.resPCA}
\title{plot.resPCA}
\usage{
\method{plot}{resPCA}(respca, type = "ind", axes = c(1, 2))
}
\arguments{
\item{respca}{result of a PCA function. See \link{PCA}}

\item{type}{ind or var}

\item{axes}{vector containing the two axes to be plotted. Default to c(1,2)}
}
\description{
plot.resPCA
}
