% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCVAgg.r
\name{plotCVAgg}
\alias{plotCVAgg}
\title{plotCVAgg
Plots the results of a CVA}
\usage{
plotCVAgg(
  respcagg,
  type = "ind",
  text = TRUE,
  n = 10,
  colorInd = "all",
  substrVec = c(1, 2),
  axes = c(1, 2),
  indSup = c("ell"),
  repel = FALSE,
  revertX = FALSE,
  revertY = FALSE,
  sizeText = NULL,
  segmentsHotelling = TRUE
)
}
\arguments{
\item{respcagg}{res of PCA gg}

\item{type}{"ind" for individual graph,"distanceBiplot" for biplot,  "var" for variable graph,"cor" for univariate correlations in decreasing order with the PCA axes}

\item{text}{Boolean indicating whether the labels should be displayed (TRUE) or not (FALSE)}

\item{n}{number of variables to be selected in correlation graph (when type="cor")}

\item{colorInd}{if type(respcagg)="raw", this parameter allows to color the individuals according to a letter in the name of individual. It can be the first one (with "first") or another subset of character ("substr") defined in substrVec}

\item{substrVec}{vector of two numbers representing the two limit positions of characters to be selected when colorInd="substr"}

\item{axes}{c(1,2). Axes to be plotted}

\item{indSup}{"ell","points" or "none"}

\item{repel}{if TRUE, ggrepel is used to place the labels}

\item{revertX}{if TRUE the xaxis is reverted}

\item{revertY}{if TRUE the yaxis is reverted}

\item{sizeText}{number allowing to change the size of the text}
}
\description{
plotCVAgg
Plots the results of a CVA
}
