% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starPValue.R
\name{starPValue}
\alias{starPValue}
\title{Get stars from a gicven pvalue.}
\usage{
starPValue(x)
}
\arguments{
\item{x}{Numeric. Vector of p values.}
}
\value{
Character.
}
\description{
Get stars from a gicven pvalue.
}
\keyword{internal}
