% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnToPcagg.r
\name{turnToPCAgg}
\alias{turnToPCAgg}
\title{turnToPCAgg
Return a pcagg object from a PCA object}
\usage{
turnToPCAgg(
  resPca,
  axes = list(c(1, 2)),
  representation = "DistanceBiplot",
  expandBiplot = NULL
)
}
\arguments{
\item{resPca}{result of PCA function}

\item{axes}{the two axes to be chosen for the plot}

\item{representation}{"DistanceBiplot","CorrelationBiplot" or "TwoMaps".}

\item{expandBiplot}{numeric that allows to custumize the size of variables arrows. If NULL, this parameter is optimized according to the graph.}
}
\value{
listCoord
}
\description{
turnToPCAgg
Return a pcagg object from a PCA object
}
