% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotellingTest.R
\name{HotellingTest}
\alias{HotellingTest}
\title{HotellingTest
analyse p variee : mat x et mat y auront p colonnes}
\usage{
HotellingTest(matx, maty)
}
\description{
HotellingTest
analyse p variee : mat x et mat y auront p colonnes
}
