% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAgg.r
\name{PCAgg}
\alias{PCAgg}
\title{PCAgg
Performs a PCA from a long dataset (potentially by rep/period or rep and period)}
\usage{
PCAgg(
  df,
  dataType = "productMeans",
  option = "Covariance",
  representation = "DistanceBiplot",
  value.var = "score",
  variable = "descriptor",
  axes = list(c(1, 2)),
  expandBiplot = NULL,
  confInt = 0.9,
  ellipseType = "barycentric",
  ellipseCalculation = "Chi",
  bootstrap = FALSE,
  nSamples = 100
)
}
\arguments{
\item{df}{with subject, product, descriptor, score, rep and period}

\item{dataType}{"productMeans" or "raw"}

\item{option}{"Covariance" or "Correlations".}

\item{representation}{"DistanceBiplot","CorrelationBiplot" or "TwoMaps".}

\item{value.var}{Name of the score column ("score" by default)}

\item{variable}{Name of the variable column ("descriptor" by default)}

\item{axes}{the two axes to be chosen for the plot}

\item{expandBiplot}{numeric that allows to custumize the size of variables arrows. If NULL, this parameter is optimized according to the graph.}

\item{confInt}{threshold for confidence intervall. A value of 0.9 (the default) for barycentric ellipses indicates that the mean is inside the ellipse at 90\%. For individual ellipses, one individual has 90\% of chances to be in the ellipse}

\item{ellipseType}{"barycentric" or "individual". Barycentric is the confidence ellipse for the mean while individual is the confidence ellipse for the individuals.}

\item{ellipseCalculation}{"Chi","F" or "SAS". Method to estimate the ellipse calculation.}

\item{bootstrap}{=FALSE If TRUE the ellipses represents the variability}

\item{nSamples}{=100 Numer of samples used with bootstrap option}
}
\value{
listCoord
}
\description{
PCAgg
Performs a PCA from a long dataset (potentially by rep/period or rep and period)
}
